Inject prelude library into the compiler itself
This speeds up type checking as the type checker doesn't need to consult
the runtime jar every time it uses prelude module. However, to get the
information right, we need to do a little bit more if anything changes
in the prelude module. Below is the scenario:

1. Someone modifies the prelude module

2. To let the compiler know the changes, we first recompile the prelude
module using the old compiler, and get some Java files

3. Then we rebuild the runtime jar with the newly generated Java files

4. Finally we rebuild the compiler. During building, GHC grabs the
information from the new runtime jar, and injects the prelude module
into the executable

Refactor the testing module, and remove some kludge